perf-report: remove some nonsensical reports
Since it is be possible for events to be processed while
we are trying to print a report, we could end in a situation where the
report contains events that happened _after_ the `now` timestamp.
This would lead to nonsensical reports.  As such now each thread gets
its own timestamp such that no events were processed after that.