jcr: split creating jcr with registering jcr
Because the jcr is not fully created after new_jcr (the daemon
specific pointer is still null), it is not 100% safe to register it
since a lot of jcr-chain-walkers expect that pointer to be set when
they walk the chain.
This introduces a new register_jcr function which handles just the
registering so that it can be done after the initialisation is
actually over.